oncoped/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
arrow==1.2.2
    # via
    #   -r requirements.txt
    #   django-q
asgiref==3.5.0
    # via
    #   -r requirements.txt
    #   django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
black==22.1.0
    # via -r requirements-dev.in
blessed==1.19.1
    # via
    #   -r requirements.txt
    #   django-q
click==8.0.4
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via
    #   -r requirements.txt
    #   odfpy
diff-match-patch==20200713
    # via
    #   -r requirements.txt
    #   django-import-export
dj-database-url==0.5.0
    # via -r requirements.txt
django==3.2.12
    # via
    #   -r requirements.txt
    #   django-debug-toolbar
    #   django-extensions
    #   django-import-export
    #   django-jazzmin
    #   django-js-asset
    #   django-multiselectfield
    #   django-picklefield
    #   django-q
    #   django-recaptcha
    #   django-storages
django-ckeditor==6.2.0
    # via -r requirements.txt
django-crispy-forms==1.14.0
    # via -r requirements.txt
django-debug-toolbar==3.2.4
    # via -r requirements-dev.in
django-environ==0.8.1
    # via -r requirements.txt
django-extensions==3.1.5
    # via -r requirements-dev.in
django-import-export==2.7.1
    # via -r requirements.txt
django-jazzmin==2.4.9
    # via -r requirements.txt
django-js-asset==2.0.0
    # via
    #   -r requirements.txt
    #   django-ckeditor
django-multiselectfield==0.1.12
    # via -r requirements.txt
django-picklefield==3.0.1
    # via
    #   -r requirements.txt
    #   django-q
django-q==1.3.9
    # via
    #   -r requirements.txt
    #   django-q-email
django-q-email==5.0.1
    # via -r requirements.txt
django-recaptcha==3.0.0
    # via -r requirements.txt
django-storages==1.12.3
    # via -r requirements.txt
et-xmlfile==1.1.0
    # via
    #   -r requirements.txt
    #   openpyxl
executing==0.8.3
    # via stack-data
fancycompleter==0.9.1
    # via pdbpp
gevent==21.12.0
    # via -r requirements.txt
greenlet==1.1.2
    # via
    #   -r requirements.txt
    #   gevent
gunicorn==20.1.0
    # via -r requirements.txt
ipython==8.1.1
    # via -r requirements-dev.in
isort==5.10.1
    # via -r requirements-dev.in
jedi==0.18.1
    # via ipython
markuppy==1.14
    # via
    #   -r requirements.txt
    #   tablib
matplotlib-inline==0.1.3
    # via ipython
mypy-extensions==0.4.3
    # via black
numpy==1.22.3
    # via
    #   -r requirements.txt
    #   pandas
odfpy==1.4.1
    # via
    #   -r requirements.txt
    #   tablib
openpyxl==3.0.9
    # via
    #   -r requirements.txt
    #   tablib
pandas==1.4.1
    # via -r requirements.txt
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pdbpp==0.10.3
    # via -r requirements-dev.in
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.5.1
    # via -r requirements-dev.in
platformdirs==2.5.1
    # via black
prompt-toolkit==3.0.28
    # via ipython
psycopg2==2.9.3
    # via -r requirements.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.11.2
    # via
    #   ipython
    #   pdbpp
pymemcache==3.5.1
    # via -r requirements.txt
pyrepl==0.9.0
    # via fancycompleter
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   arrow
    #   pandas
pytz==2021.3
    # via
    #   -r requirements.txt
    #   django
    #   pandas
pyyaml==6.0
    # via
    #   -r requirements.txt
    #   tablib
redis==3.5.3
    # via
    #   -r requirements.txt
    #   django-q
six==1.16.0
    # via
    #   -r requirements.txt
    #   blessed
    #   pymemcache
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   -r requirements.txt
    #   django
    #   django-debug-toolbar
stack-data==0.2.0
    # via ipython
supervisor==4.2.4
    # via -r requirements.txt
tablib[html,ods,xls,xlsx,yaml]==3.2.0
    # via
    #   -r requirements.txt
    #   django-import-export
tomli==2.0.1
    # via
    #   black
    #   pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.1.1
    # via black
wcwidth==0.2.5
    # via
    #   -r requirements.txt
    #   blessed
    #   prompt-toolkit
werkzeug==2.0.3
    # via -r requirements-dev.in
wheel==0.37.1
    # via pip-tools
whitenoise==6.0.0
    # via -r requirements.txt
wmctrl==0.4
    # via pdbpp
xlrd==2.0.1
    # via
    #   -r requirements.txt
    #   tablib
xlsxwriter==3.0.3
    # via -r requirements.txt
xlwt==1.3.0
    # via
    #   -r requirements.txt
    #   tablib
zope-event==4.5.0
    # via
    #   -r requirements.txt
    #   gevent
zope-interface==5.4.0
    # via
    #   -r requirements.txt
    #   gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
